Fix conversion error
```bash
gcc  -o yk_demo_hash yk_demo_hash.c ../lib/massdal/prng.c -ltrace -lm
../lib/massdal/prng.c: In function ‘RanrotAInit’:
../lib/massdal/prng.c:209:26: warning: overflow in conversion from ‘long unsigned int’ to ‘int’ changes value from ‘18446744073709551552’ to ‘-64’ [-Woverflow]
  209 |   prng->scale = ldexp(1, -8*sizeof(unsigned long));
      |                          ^~~~~~~~~~~~~~~~~~~~~~~~
```

+++ lib/massdal/prng.c
@@ -206,7 +206,7 @@ void RanrotAInit (prng_type * prng, unsigned long seed) {
 
   /* randomize */
   for (i = 0;  i < 300;  i++) ran3(prng);
-  prng->scale = ldexp(1, -8*sizeof(unsigned long));
+  prng->scale = ldexp(1, -8*(long)sizeof(unsigned long));
 }
